fix warning CS860-CS864

+++ DistributionOfStudents/Controllers/AdmissionsController.cs
@@ -38,8 +38,8 @@ namespace DistributionOfStudents.Controllers
         // GET: Admissions/Details/5
         public async Task<IActionResult> Details(string facultyName, int groupId, int id)
         {
-           
-            Admission admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
+
+            Admission? admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
 
             if (admission == null)
             {
@@ -53,13 +53,17 @@ namespace DistributionOfStudents.Controllers
         // GET: Admissions/Create
         public async Task<IActionResult> Create(string facultyName, int groupId)
         {
-            GroupOfSpecialties group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
+            GroupOfSpecialties? group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
+            if (group == null)
+            {
+                return NotFound();
+            }
             List<StudentScore> scores = new();
             List<SpecialityPriorityVM> priorities = new();
             List<RecruitmentPlan> plans = await _planRepository.GetAllAsync(new RecruitmentPlansSpecification().WhereFaculty(facultyName).WhereGroup(group));
-            plans = plans.Where(p => group.Specialities.Contains(p.Speciality)).ToList();
+            plans = plans.Where(p => (group.Specialities ?? new()).Contains(p.Speciality)).ToList();
 
-            foreach (Subject subject in group.Subjects)
+            foreach (Subject subject in group.Subjects ?? new())
             {
                 StudentScore score = new()
                 {
@@ -101,18 +105,28 @@ namespace DistributionOfStudents.Controllers
 
                 foreach (StudentScore studentScore in model.StudentScores)
                 {
-                    Subject subject = await _subjectsRepository.GetByIdAsync(studentScore.Subject.Id);
-                    studentScore.Subject = subject;
+                    if (studentScore.Subject != null)
+                    {
+                        Subject? subject = await _subjectsRepository.GetByIdAsync(studentScore.Subject.Id);
+                        if (subject != null)
+                        {
+                            studentScore.Subject = subject;
+                        }
+                    }
                 }
                 foreach (SpecialityPriorityVM specialityPriority in model.SpecialitiesPriority.Where(p => p.Priority > 0))
                 {
-                    SpecialityPriority priority = new() { Priority = specialityPriority.Priority, RecruitmentPlan = await _planRepository.GetByIdAsync(specialityPriority.PlanId) };
-                    specialityPriorities.Add(priority);
+                    RecruitmentPlan? plan = await _planRepository.GetByIdAsync(specialityPriority.PlanId);
+                    if (plan != null)
+                    {
+                        SpecialityPriority priority = new() { Priority = specialityPriority.Priority, RecruitmentPlan = plan };
+                        specialityPriorities.Add(priority);
+                    }
                 }
 
                 Admission admission = new()
                 {
-                    GroupOfSpecialties = await _groupsRepository.GetByIdAsync(groupId),
+                    GroupOfSpecialties = await _groupsRepository.GetByIdAsync(groupId) ?? new(),
                     StudentScores = model.StudentScores,
                     Student = new() { Name = model.Student.Name, Surname = model.Student.Surname, Patronymic = model.Student.Patronymic, GPS = model.Student.GPS },
                     DateOfApplication = model.DateOfApplication,
@@ -131,16 +145,16 @@ namespace DistributionOfStudents.Controllers
         // GET: Admissions/Edit/5
         public async Task<IActionResult> Edit(string facultyName, int groupId, int id)
         {
-            Admission admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
-            if (admission == null)
+            Admission? admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
+            GroupOfSpecialties? group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
+            if (admission == null || group == null)
             {
                 return NotFound();
             }
 
-            GroupOfSpecialties group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
             List<SpecialityPriorityVM> priorities = new();
             List<RecruitmentPlan> plans = await _planRepository.GetAllAsync(new RecruitmentPlansSpecification().WhereFaculty(facultyName).WhereGroup(group));
-            plans = plans.Where(p => group.Specialities.Contains(p.Speciality)).ToList();
+            plans = plans.Where(p => (group.Specialities ?? new()).Contains(p.Speciality)).ToList();
 
             foreach (RecruitmentPlan plan in plans)
             {
@@ -180,7 +194,11 @@ namespace DistributionOfStudents.Controllers
         {
             if (ModelState.IsValid)
             {
-                Admission admission = await _admissionsRepository.GetByIdAsync(model.Id ?? 0, new AdmissionsSpecification().IncludeSpecialtyPriorities().IncludeStudentScores());
+                Admission? admission = await _admissionsRepository.GetByIdAsync(model.Id ?? 0, new AdmissionsSpecification().IncludeSpecialtyPriorities().IncludeStudentScores());
+                if (admission == null)
+                {
+                    return RedirectToAction("Details", "GroupsOfSpecialties", new { facultyName, id = groupId });
+                }
 
                 List<SpecialityPriority> specialityPriorities = new();
                 List<Subject> subjects = new();
@@ -192,7 +210,7 @@ namespace DistributionOfStudents.Controllers
                 foreach (SpecialityPriorityVM specialityPriority in model.SpecialitiesPriority.Where(p => p.Priority > 0))
                 {
                     SpecialityPriority priority = admission.SpecialityPriorities.FirstOrDefault(i => i.RecruitmentPlan.Id == specialityPriority.PlanId) ??
-                        new() { RecruitmentPlan = await _planRepository.GetByIdAsync(specialityPriority.PlanId) };
+                        new() { RecruitmentPlan = await _planRepository.GetByIdAsync(specialityPriority.PlanId) ?? new() };
                     priority.Priority = specialityPriority.Priority;
                     specialityPriorities.Add(priority);
                 }
@@ -228,9 +246,12 @@ namespace DistributionOfStudents.Controllers
 
         public async Task<IActionResult> Delete(string facultyName, int groupId, int id)
         {
-            Admission admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification());
-            await _admissionsRepository.DeleteAsync(id);
-            _logger.LogInformation("Заявка студента - {Surname} {Name} {Patronymic} - была удалена", admission.Student.Surname, admission.Student.Name, admission.Student.Patronymic);
+            Admission? admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification());
+            if (admission != null)
+            {
+                await _admissionsRepository.DeleteAsync(id);
+                _logger.LogInformation("Заявка студента - {Surname} {Name} {Patronymic} - была удалена", admission.Student.Surname, admission.Student.Name, admission.Student.Patronymic);
+            }
 
             return RedirectToAction("Details", "GroupsOfSpecialties", new { facultyName, id = groupId });
         }
